# float32 hit pos, faster simd unpackAlphaValues