--- a/src/pixie/blends.nim
+++ b/src/pixie/blends.nim
@@ -166,9 +166,7 @@ proc SetSat(C: Color, s: float32): Color {.inline.} =
     result = (C - min([C.r, C.g, C.b])) * s / satC
 
 proc blendNormal(backdrop, source: ColorRGBX): ColorRGBX =
-  if backdrop.a == 0:
-    return source
-  if source.a == 255:
+  if backdrop.a == 0 or source.a == 255:
     return source
   if source.a == 0:
     return backdrop