… examples/heart.png
… examples/masking.png
… examples/text.png
… examples/text_spans.png
… examples/tiger.png
… tests/contexts/quadracticCurveTo_2.png
… tests/fileformats/svg/diffs/Ghostscript_Tiger.png
… tests/fileformats/svg/diffs/accessibility-outline.png
… tests/fileformats/svg/diffs/circle01.png
… tests/fileformats/svg/diffs/ellipse01.png
… tests/fileformats/svg/diffs/miterlimit.png
… tests/fileformats/svg/diffs/polygon01.png
… tests/fileformats/svg/diffs/quad01.png
… tests/fileformats/svg/diffs/rect02.png
… tests/fileformats/svg/emojitwo.png
… tests/fileformats/svg/flat-color-icons.png
… tests/fileformats/svg/ionicons.png
… tests/fileformats/svg/noto-emoji.png
… tests/fileformats/svg/openmoji.png
… tests/fileformats/svg/rendered/Ghostscript_Tiger.png
… tests/fileformats/svg/rendered/accessibility-outline.png
… tests/fileformats/svg/rendered/circle01.png
… tests/fileformats/svg/rendered/ellipse01.png
… tests/fileformats/svg/rendered/miterlimit.png
… tests/fileformats/svg/rendered/polygon01.png
… tests/fileformats/svg/rendered/quad01.png
… tests/fileformats/svg/rendered/rect02.png
… tests/fileformats/svg/simple-icons.png
… tests/fileformats/svg/tabler-icons.png
… tests/fileformats/svg/twbs-icons.png