does not result in leaks per leak fuzzing tests

before:
`newImage(mask) ..................... 1.216 ms      1.246 ms    ±0.040   x1000`
`svg decode ........................ 14.268 ms     14.358 ms    ±0.108   x348`
`fill text .......................... 2.959 ms      3.004 ms    ±0.037  x1000`

after:
`newImage(mask) ..................... 0.828 ms      0.845 ms    ±0.023   x1000`
`svg decode ........................ 13.982 ms     14.053 ms    ±0.076   x356`
`fill text .......................... 2.955 ms      2.986 ms    ±0.027  x1000`